Create const_cast
Write a program that prompts the user to enter a pointer to an object of a certain class and then converts it to a pointer to an object of a different class using the const_cast operator. The program should then display the original pointer and the converted pointer to the user.